[GEN-1815] Refactor coin selection
