+++ src/Helpers/LightningHelper.cs
@@ -133,94 +133,6 @@ namespace NodeGuard.Helpers
         }
 
         /// <summary>
-        /// Helper to select utxos from a wallet for requests (Withdrawals, ChannelOperationRequest) by oldest
-        /// </summary>
-        /// <param name="wallet"></param>
-        /// <param name="satsAmount"></param>
-        /// <param name="availableUTXOs"></param>
-        /// <param name="logger"></param>
-        /// <returns></returns>
-        public static async Task<List<UTXO>> SelectUTXOsByOldest(
-            Wallet wallet, long satsAmount, List<UTXO> availableUTXOs, ILogger logger)
-        {
-            if (wallet == null) throw new ArgumentNullException(nameof(wallet));
-            if (logger == null) throw new ArgumentNullException(nameof(logger));
-            if (wallet == null) throw new ArgumentNullException(nameof(wallet));
-            if (satsAmount <= 0) throw new ArgumentOutOfRangeException(nameof(satsAmount));
-
-            var selectedUTXOs = new List<UTXO>();
-
-            if (!availableUTXOs.Any())
-            {
-                logger.LogError("The PSBT cannot be generated, no UTXOs are available for walletId: {WalletId}",
-                    wallet.Id);
-                return selectedUTXOs;
-            }
-
-            var utxosStack = new Stack<UTXO>(availableUTXOs.OrderByDescending(x => x.Confirmations));
-
-            //FIFO Algorithm to match the amount, oldest UTXOs are first taken
-
-            var totalUTXOsConfirmedSats = utxosStack.Sum(x => ((Money)x.Value).Satoshi);
-
-            if (totalUTXOsConfirmedSats < satsAmount)
-            {
-                logger.LogError(
-                    "Error, the total UTXOs set balance for walletid: {WalletId} ({AvailableSats} sats) is less than the amount in the request ({RequestedSats} sats)",
-                    wallet.Id, totalUTXOsConfirmedSats, satsAmount);
-                return selectedUTXOs;
-            }
-
-            var utxosSatsAmountAccumulator = 0M;
-
-            var iterations = 0;
-            while (satsAmount >= utxosSatsAmountAccumulator)
-            {
-                if (utxosStack.TryPop(out var utxo))
-                {
-                    selectedUTXOs.Add(utxo);
-                    utxosSatsAmountAccumulator += ((Money)utxo.Value).Satoshi;
-                }
-
-                iterations++;
-
-                if (iterations == 1_000)
-                {
-                    break;
-                }
-            }
-
-            return selectedUTXOs;
-        }
-
-        /// <summary>
-        /// Helper to select coins from a wallet for requests (Withdrawals, ChannelOperationRequest). FIFO is the coin selection
-        /// </summary>
-        /// <param name="wallet"></param>
-        /// <param name="selectedUTXOs"></param>
-        /// <returns></returns>
-        public static async Task<List<ICoin>> SelectCoins(Wallet wallet, List<UTXO> selectedUTXOs)
-        {
-            if (wallet == null) throw new ArgumentNullException(nameof(wallet));
-            if (wallet == null) throw new ArgumentNullException(nameof(wallet));
-
-            var derivationStrategy = wallet.GetDerivationStrategy();
-
-            //UTXOS to Enumerable of ICOINS
-            return selectedUTXOs.Select<UTXO, ICoin>(x =>
-                {
-                    var coin = x.AsCoin(derivationStrategy);
-                    if (wallet.IsHotWallet)
-                    {
-                        return coin;
-                    }
-
-                    return coin.ToScriptCoin(x.ScriptPubKey);
-                })
-                .ToList();
-        }
-
-        /// <summary>
         /// Returns the fee rate (sat/vb) for a tx, 4 is the value in regtest
         /// </summary>
         /// <param name="nbXplorerNetwork"></param>
